api/requirements.txt
python == 3.12.3
asgiref == 3.8.1
Django == 5.1.2
djangorestframework == 3.15.2
pip == 24.0
sqlparse == 0.5.1
tzdata == 2024.2
drf-yasg == 1.21.8
